backend/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
pandas==2.2.3
numpy
nltk
torch
faiss-cpu
gensim
pyarrow
pydantic-settings
pydantic==2.9.2
python-dotenv
minio